% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_tabby.R
\name{big_tabby}
\alias{big_tabby}
\title{Create large-scale crosstab}
\usage{
big_tabby(
  df,
  row.vars,
  col.vars = NULL,
  wts = NULL,
  split.vars = FALSE,
  percent = FALSE,
  decimal = FALSE
)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{row.vars}{A \code{character} or {character} vector}

\item{col.vars}{A \code{character} or {character} vector}

\item{wts}{A \code{character} or {character} vector}

\item{split.vars}{TRUE/FALSE: Should col.vars be treated as combined or crossed variables? Defaults to \code{FALSE}}

\item{percent}{TRUE/FALSE: Should output format be percent? Defaults to \code{FALSE}}

\item{decimal}{TRUE/FALSE: Should output format be decimal? Defaults to \code{FALSE}}
}
\value{
A \code{data.frame} with a "Question" column for the row variable names, a "Response" column for the row variable levels, columns displaying
percentages by total and grouped by column variables supplied by user.
}
\description{
Create large-scale crosstab
}
