% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_vars.R
\name{cross_vars}
\alias{cross_vars}
\title{Combine two variables for analysis by groups}
\usage{
cross_vars(df, var1, var2)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{var1}{A \code{character} indicating the name of the first variable to combine}

\item{var2}{A \code{character} indicating the name of the second variable to combine}
}
\value{

}
\description{
Combine two variables for analysis by groups
}
\examples{
# Add new combined variable to data.frame
library(tidyverse)
nationscape_2020_excerpt <- nationscape_2020_excerpt \%>\% mutate(cross_vars(., "gender", "vote_2016"))
}
