% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_factor.R
\name{df_to_factor}
\alias{df_to_factor}
\title{Transform all variables of data.frame to factors}
\usage{
df_to_factor(df)
}
\arguments{
\item{df}{A \code{data.frame}}
}
\value{
A \code{data.frame} with factors as variables
}
\description{
Transform all variables of data.frame to factors
}
\examples{
df_to_factor(test_df)
}
