% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fct_info.R
\name{get_fct_info}
\alias{get_fct_info}
\title{Get names and levels of all variables in a factored data.frame}
\usage{
get_fct_info(df_factored, row.vars)
}
\arguments{
\item{df_factored}{A \code{data.frame} where all variables have been transformed as factors.}

\item{row.vars}{A \code{character} vector}
}
\value{
A \code{list} of the names and levels of variables in row.vars
}
\description{
Get names and levels of all variables in a factored data.frame
}
\examples{
get_fct_info(factor_data, analysis_variables)
}
