% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_lengths.R
\name{get_input_lengths}
\alias{get_input_lengths}
\title{Get number of levels of factor variables in data.frame}
\usage{
get_input_lengths(df, input1, input2 = NULL)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{input1}{A \code{character} vector}

\item{input2}{An optional \code{character} vector}
}
\value{

}
\description{
Get number of levels of factor variables in data.frame
}
\examples{
get_input_lengths(tbl, row.variables, column.variables)
}
