% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacktab.R
\name{stacktab}
\alias{stacktab}
\title{Combine list of data.frames vertically as one data.frame}
\usage{
stacktab(df_list, vars)
}
\arguments{
\item{df_list}{A \code{list} of data.frames}

\item{row.vars}{A \code{character} vector}
}
\value{
A \code{data.frame}
}
\description{
Combine list of data.frames vertically as one data.frame
}
