% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_tabby.R
\name{big_tabby}
\alias{big_tabby}
\title{Create large-scale crosstab}
\usage{
big_tabby(
  df,
  row.vars,
  col.vars = NULL,
  wts = NULL,
  data.frame = FALSE,
  format = FALSE,
  split.vars = FALSE
)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{row.vars}{A \code{character} or {character} vector}

\item{col.vars}{A \code{character} or {character} vector}

\item{wts}{A \code{character} or {character} vector}

\item{data.frame}{TRUE/FALSE: Should output be returned as a data.frame? Defaults to \code{FALSE}}

\item{format}{A \code{character} indicating "percent" or "decimal" format for the totals of the variables tabulated. Only available when data.frame=TRUE.}

\item{split.vars}{TRUE/FALSE: Should col.vars be treated as combined or crossed variables? Defaults to \code{FALSE}}
}
\value{
A \code{data.frame} with a "Question" column for the row variable names, a "Response" column for the row variable levels, columns displaying
percentages by total and grouped by column variables supplied by user.
}
\description{
Create large-scale crosstab
}
\examples{
# Basic unweighted topline crosstab
big_tabby(df, row.vars)

# Weighted crosstab grouped by column variables
big_tabby(df, row.vars, col.vars, wts = "weights")
}
