% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacktab.R
\name{stacktab}
\alias{stacktab}
\title{Combine list of data.frames vertically as one data.frame}
\usage{
stacktab(x)
}
\arguments{
\item{x}{A \code{list} of data.frames}
}
\value{
A \code{data.frame}
}
\description{
Combine list of data.frames vertically as one data.frame
}
